refactor(event-handler): extract shared IP-extraction util for HTTP middleware
Move the duplicated client-IP extraction logic out of the tracer and
metrics middleware into a shared commons module, and have both import it.

Closes #5345
