+++ packages/event-handler/src/http/middleware/tracer.ts
@@ -6,6 +6,7 @@ import type {
   SegmentHttpData,
   TracerOptions,
 } from '../../types/http.js';
+import { getClientIp } from './commons.js';
 import type { compress } from './compress.js';
 
 /**
@@ -16,28 +17,6 @@ import type { compress } from './compress.js';
  * optional property to attach request/response data without unsafe casts.
  */
 type HttpSubsegment = Subsegment & { http?: SegmentHttpData };
-
-/**
- * Extracts the client IP address from the request context.
- *
- * For API Gateway events the source IP is taken from the request context, while
- * for ALB (or any other source) it falls back to the `X-Forwarded-For` header.
- *
- * @param reqCtx - The request context for the current request
- */
-const getClientIp = (reqCtx: RequestContext): string | undefined => {
-  if (reqCtx.responseType === 'ApiGatewayV1') {
-    return reqCtx.event.requestContext.identity.sourceIp;
-  }
-  if (reqCtx.responseType === 'ApiGatewayV2') {
-    return reqCtx.event.requestContext.http.sourceIp;
-  }
-  const xForwardedFor = reqCtx.req.headers.get('X-Forwarded-For');
-  if (xForwardedFor) {
-    return xForwardedFor.split(',')[0].trim();
-  }
-  return undefined;
-};
 
 /**
  * Builds the `http.request` data for the X-Ray subsegment from the request context.
